Code reorg and fixup the missing piece of Mapper
